+++ CHANGES.rst
@@ -1,7 +1,7 @@
 0.1.29 (unreleased)
 -----------------------
 
-- No changes yet.
+- Added decorator to loop over different set of values (:py:func:`~cosmo_utils.utils.file_utils.mark_parametrize`).
 
 0.1.28 (2018-05-21)
 -----------------------